requirements.txt
pyodbc
requests
python-dotenv
jira
colorlog
odfpy
pyexcel-ods
pyexcel-ods3
reportlab
unotools
pyuno